Python-Learning/Python-Django/Python-Coding-Challenge/Python-List/Advance/Map_Locater/requirement.txt
geopy-2.4.1
folium